scripts/requirements_mcp.txt
mcp
paddlepaddle
paddleocr
pdf2image
opencv-python
numpy
